fix(client): avoid exporting internal catalog uniqueness macro from published crate (Codex review)
- Remove unique-macro.inc.rs + include sharing that caused
  #[macro_export] on __capability_catalog_unique_impl! to leak into
  the publishable openlark-client crate (P1 hard violation).
- Keep full macro bodies in client (private, no export) and harness
  (exported only for trybuild) to preserve publish isolation.
- The two implementations remain text-identical for consistency.
- Clean trailing blank line in allowlist (git diff --check).
- Revert unrelated crates.md API count update (divergent change).
- Update comments; bless updated stderr expectations.

Addresses latest Codex review on PR #455 / #423:
- No internal macros exposed from client.
- No unrelated changes mixed in.
- Production happy-path + trybuild negative tests still cover
  generation-time duplicate rejection.

All tests, clippy and fmt checks pass.
