+++ tools/mod_reachability_allowlist.txt
@@ -3,4 +3,3 @@
 # CI 只对【新增】孤儿失败；各 crate 修复死代码后从此文件删除对应行，
 # 再跑 --update-allowlist 或直接编辑收敛。
 # 当前存量孤儿：0 个
-
